offer page components
