+++ src/pages/offer-list/OfferStatusBadge.svelte
